nspn500 atlas

we designed a novel approach aimed at determining the best trade-off between small-world attributes of anatomical cortical networks and the number of cortical regions included in the scale. Results revealed that schemes comprising 540-599 regions (surface areas spanning between 250 and 275 mm(2)) at sparsities below 10 a superior balance between small-world organization and the size of the cortical scale employed. annot files were acquired from (faskowit/multiAtlasTT)

nspn500

nspn500_3d

Format

An object of class brain_atlas of length 4.

An object of class ggseg3d_atlas (inherits from tbl_df, tbl, data.frame) with 4 rows and 4 columns.

References

Whitaker, K. J., Vértes, P. E., Romero-Garcia, R., Váša, F., Moutoussis, M., Prabhu, G., ... & Tait, R. (2016). Adolescence is associated with genomically patterned consolidation of the hubs of the human brain connectome. Proceedings of the National Academy of Sciences, 113(32), 9105-9110. Romero-Garcia, R., Atienza, M., Clemmensen, L. H., & Cantero, J. L. (2012). Effects of network resolution on topological properties of human neocortex. Neuroimage, 59(4), 3522-3532. (PubMed)

nspn500 - nspn500 atlas

nspn500_3d - nspn500 3d mesh atlas

Examples

data(nspn500)
data(nspn500_3d)